wip request with pagination
